requirements.txt
pandas
numpy
librosa
soundfile
datasets
huggingface_hub
torch
torchaudio
pydub
scipy
tqdm
gTTS
edge-tts
scikit-learn
xgboost
joblib
fastapi
uvicorn
python-multipart
transformers
gradio
speechbrain
python-dotenv
pydantic
httpx
matplotlib
